add other bot page

+++ linebot/getIdToken/views.py
@@ -1,82 +1,107 @@
-from rest_framework import status
-from django.shortcuts import render
-from django.http import HttpResponse
 import json
-from .models import Bot
-from .models import Activation
-from django.db.models import Q
 from datetime import datetime
+
+import requests
 from django.core import serializers
+from django.db.models import Q
 from django.forms.models import model_to_dict
+from django.http import HttpResponse
+from django.shortcuts import render
+from rest_framework import status
+
+from .auth import verify_user
+from .models import Activation, Bot, User
+from .serializer import BotSerializer
+
 
 def index(request):
     return HttpResponse("Hello, world. You're at the getIdToken index.")
 
+
 def getIdToken(request):
-    #/login/getIdToken?
-    if request.method == 'POST':
+    if request.method == "POST":
         if request.POST:
-            idToken = request.POST.get('idToken', 0)
-            return HttpResponse('Good!')
+            idToken = request.POST.get("idToken", 0)
+            return HttpResponse("Good!")
         else:
-            return HttpResponse('ERROR: It is a wrong input.')
+            return HttpResponse("ERROR: It is a wrong input.")
+
 
 def others_bot(request):
-    if request.method == 'GET':
-        developerId = request.GET.get('developerId')
-        bots = Bot.objects.filter(is_public=True).filter(~Q(developerId=developerId))
-        return HttpResponse(serializers.serialize('json', bots))
+    if request.method == "GET":
+        user = verify_user(request)
+        if user is None:
+            return HttpResponse("Auth failed", status=403)
+        bots = Bot.objects.filter(is_public=True).order_by("-updated_at")
+        serializer = BotSerializer(bots, many=True)
+        return HttpResponse(json.dumps(serializer.data))
 
-    return HttpResponse('Only GET requests are allowed')
+    return HttpResponse("Only GET requests are allowed")
+
 
 def root(request):
-    developerId = request.GET.get('developerId')
-    print(developerId)
-    if request.method == 'GET':
-        return index(request, developerId)
-    elif request.method == 'POST':
+    if request.method == "GET":
+        user = verify_user(request)
+        if user is None:
+            return HttpResponse("Auth failed", status=403)
+        bots = Bot.objects.filter(created_by=user.user_id).order_by("-updated_at")
+        serializer = BotSerializer(bots, many=True)
+        return HttpResponse(json.dumps(serializer.data))
+    elif request.method == "POST":
         return create(request)
 
-    return HttpResponse('Only GET and POST requests are allowed')
+    return HttpResponse("Only GET and POST requests are allowed")
+
 
 def id(request, id):
-    if request.method == 'GET':
-        return show(request, id)
-    elif request.method == 'DELETE':
+    if request.method == "GET":
+        bot = Bot.objects.get(bot_id=id)
+
+        if not bot.is_public:
+            user = verify_user(request)
+            if user is None:
+                return HttpResponse("Unauthorized", status=401)
+            if user.user_id != bot.created_by.user_id:
+                return HttpResponse("Forbidden", status=403)
+
+        serializer = BotSerializer(bot)
+
+        return HttpResponse(json.dumps(serializer.data))
+    elif request.method == "DELETE":
         return delete(request, id)
-    elif request.method == 'PUT':
+    elif request.method == "PUT":
         return update(request, id)
 
-    return HttpResponse('Only GET and DELETE requests are allowed')
+    return HttpResponse("Only GET and DELETE requests are allowed")
 
-def index(request, developerId):
-    bots = Bot.objects.filter(developerId = developerId)
 
-    return HttpResponse(serializers.serialize('json', bots))
+def login(request):
+    if request.method == "POST":
+        user = verify_user(request)
+        if user is None:
+            return HttpResponse("Auth failed", status=403)
 
-def show(request, id):
-    bot = Bot.objects.get(bot_id=id)
-    bot_dict = model_to_dict(bot)
+        user_dict = model_to_dict(user)
+        return HttpResponse(json.dumps(user_dict))
 
-    return HttpResponse(json.dumps(bot_dict))
-    
+    return HttpResponse("Only POST requests are allowed")
+
 
 def create(request):
-    if request.method != 'POST':
-        return HttpResponse('Only POST requests are allowed')
+    if request.method != "POST":
+        return HttpResponse("Only POST requests are allowed")
 
     body = json.loads(request.body)
-
-    now = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+    user = verify_user(request)
+    if user is None:
+        return HttpResponse("Auth failed", status=403)
 
     bot = Bot(
-        bot_id=body['bot_id'],
-        name=body['name'],
-        developerId=body['developerId'],
-        flowChart=body['flowChart'],
-        is_public=body['is_public'],
-        createdAt=now,
-        updateAt=now
+        bot_id=body["bot_id"],
+        name=body["name"],
+        created_by=user,
+        flowchart=body["flowchart"],
+        is_public=False,
     )
 
     bot.save()
@@ -84,56 +109,68 @@ def create(request):
 
     return HttpResponse(json.dumps(bot_dict))
 
+
 def delete(request, id):
-    if request.method != 'DELETE':
-        return HttpResponse('Only DELETE requests are allowed')
+    if request.method != "DELETE":
+        return HttpResponse("Only DELETE requests are allowed")
+
+    user = verify_user(user)
+
+    if user is None:
+        return HttpResponse("Unauthorized", status=401)
 
     bot = Bot.objects.get(bot_id=id)
+
+    if bot.created_by.user_id != user.id:
+        return HttpResponse("Forbidden", status=403)
+
     bot.delete()
-    return HttpResponse('Bot deleted')
+    return HttpResponse("Bot deleted")
+
 
 def update(request, id):
-    if request.method != 'PUT':
-        return HttpResponse('Only PUT requests are allowed')
-    
-    bot = Bot.objects.get(bot_id=id)
-    
+    if request.method != "PUT":
+        return HttpResponse("Only PUT requests are allowed")
+
     body = json.loads(request.body)
-    
+
+    user = verify_user(request)
+    if user is None:
+        return HttpResponse("Auth failed", status=403)
+    bot = Bot.objects.get(bot_id=id)
+
     if "name" in body:
-        bot.name=body['name']
+        bot.name = body["name"]
 
-    if "developerId" in body:
-        bot.developerId=body['developerId']
+    if "created_by" in body:
+        bot.created_by = user
 
-    if "flowChart" in body:
-        bot.flowChart=body['flowChart']
-    
+    if "flowchart" in body:
+        bot.flowchart = body["flowchart"]
+
     if "is_public" in body:
-        bot.is_public=body['is_public']
-
-    bot.updateAt=datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+        bot.is_public = body["is_public"]
 
     bot.save()
     bot_dict = model_to_dict(bot)
 
     return HttpResponse(json.dumps(bot_dict))
 
+
 def activate(request):
-    if request.method != 'POST':
-        return HttpResponse('Only DELETE requests are allowed')
+    if request.method != "POST":
+        return HttpResponse("Only DELETE requests are allowed")
 
     body = json.loads(request.body)
 
     _, created = Activation.objects.update_or_create(
-      
-      user_id=body['user_id'],
-      defaults={
-        "bot_id": body['bot_id'],
-      }
+        user_id=body["user_id"],
+        defaults={
+            "bot_id": body["bot_id"],
+        },
     )
 
     if created:
-      return HttpResponse("Activated")
+        return HttpResponse("Activated")
     else:
-      return HttpResponse("Activated", status=status.HTTP_409_CONFLICT)
+        return HttpResponse("Activated", status=status.HTTP_409_CONFLICT)
